Removed equation numbers
When interacting with the Jupyter notebook, at least tables and equations are rendered differently than on GitHub. Most critically, GitHub renders equation numbers for the `align` environment which only apply to the scope of a cell, which is useless. For this reason I changed all of these to the `align*` environment, removing the equation numbers.